+++ src/main/kotlin/com/abeade/plugin/figma/ImportAction.kt
@@ -5,6 +5,7 @@ import com.android.tools.idea.rendering.webp.ConvertToWebpAction
 import com.intellij.ide.util.PropertiesComponent
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
@@ -24,15 +25,16 @@ class ImportAction : AnAction() {
 
     private lateinit var virtualFileRes: VirtualFile
 
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
+
     override fun update(anActionEvent: AnActionEvent) {
         val project = anActionEvent.project
-        val psiElement: PsiElement? = anActionEvent.dataContext.getData(PlatformDataKeys.PSI_ELEMENT)
-        val isValid = project != null && psiElement != null &&
-                (psiElement as? PsiDirectoryImpl)?.isDirectory == true &&
-                (psiElement as? PsiDirectoryImpl)?.name == RES_DIRECTORY
+        val psiElement = anActionEvent.getData(PlatformDataKeys.PSI_ELEMENT)
+        val resDirectory = psiElement as? PsiDirectoryImpl
+        val isValid = project != null && resDirectory?.isDirectory == true && resDirectory.name == RES_DIRECTORY
         anActionEvent.presentation.isEnabledAndVisible = isValid
         if (isValid) {
-            virtualFileRes = (psiElement as PsiDirectoryImpl).virtualFile
+            virtualFileRes = resDirectory!!.virtualFile
         }
     }