+++ CHANGELOG.md
@@ -14,6 +14,16 @@
 
 ### Fixed
 
+## 1.4.1 - 2025-04-30
+
+### Changed
+
+- Updated getMinSdkVersion function to avoid usage of deprecated API
+
+### Fixed
+
+- Plugin not working on AS Meerkat due to missing action update thread
+
 ## 1.4.0 - 2025-01-05
 
 ### Added